feat(S13P11D204-234) : 채팅 UI 및 테스트 API 구현

+++ bookglebookgle-android/app/src/main/java/com/ssafy/bookglebookgle/entity/ChatMessagesResponse.kt
@@ -9,7 +9,7 @@ data class ChatMessagesResponse(
 // 개별 채팅 메시지
 data class ChatMessage(
     val messageId: Long,
-    val userId: Long,
+    val userId: Int,
     val nickname: String,
     val profileImage: String?,
     val message: String,
